+++ docs/changelog.md
@@ -17,6 +17,8 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Rename `ldap` to `ldapsearch` in `sync_map`.
 - Ignore spurious ldapsearch YAML parameter.
 - Implement STARTTLS behaviour.
+- Search for .yaml file as well as .yml.
+- Drop legacy `LDAP_BIND` env var support. Use `LDAPBINDDN` instead.
 
 
 # ldap2pg 5.6